First large commit with builder, view, model, validator, setup.py
Now a lot of developpement is done on this commit.
Model View Controller pattern is adopted with a builder factory.

view.py
Is necessary to transform model to view. The tree widget items
are produced from builder factory function which takes the current
state of a model. Renders it as a tree widget.

model.py
Holds the data model that would be rendered by the view.py
The models are build from model_builder in builder factory.
However model.py should have several functions that are
defined any model in question. The model builder uses
schema for defining the model. Model.py adds functions for
setting and gettin data to models.

builder.py
Is the factory of our application. It has several purposes.
It constructs models from schemas that are given as json.
It constructs tree widget items from data underlying the dictionary.
It should also construct a path for all elements in data underlying dictionary.

validator.py
Simply checks whether the added data conforms to schema or not.

setup.py
Simply helps to install the library.

dict.json entry.json
Example schema files